User Settings Page
Code for a task with the purpose of creating a profile settings page for users where alterations could be made to their avatar, name, email, and more. I created an additional component for the task where users could see the information of their current reservation displayed, with navigation to their order history.